+++ src/main/java/com/sparta/limited/order_service/application/service/OrderServiceImpl.java
@@ -3,11 +3,11 @@ package com.sparta.limited.order_service.application.service;
 import com.sparta.limited.order_service.application.dto.request.OrderCreateRequest;
 import com.sparta.limited.order_service.application.dto.response.OrderCreateResponse;
 import com.sparta.limited.order_service.application.mapper.OrderMapper;
+import com.sparta.limited.order_service.application.service.user.UserClientService;
+import com.sparta.limited.order_service.application.service.user.UserInfo;
 import com.sparta.limited.order_service.domain.model.Order;
 import com.sparta.limited.order_service.domain.repository.OrderRepository;
 import com.sparta.limited.order_service.domain.validator.OrderValidator;
-import com.sparta.limited.order_service.infrastructure.client.dto.UserInfo;
-import com.sparta.limited.order_service.infrastructure.service.UserClientService;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
